Move ChatEventStore to :feature-chat
This also moves the koin module in there too
