+++ app/feature/feature-chat/build.gradle.kts
@@ -5,6 +5,8 @@ plugins {
 }
 
 dependencies {
+  implementation(libs.androidx.datastore.core)
+  implementation(libs.androidx.datastore.preferencesCore)
   implementation(libs.androidx.navigation.common)
   implementation(libs.androidx.navigation.compose)
   implementation(libs.androidx.navigation.runtime)
