+++ app/feature/feature-chat/src/main/kotlin/com/hedvig/android/feature/chat/di/ChatModule.kt
@@ -1,0 +1,9 @@
+package com.hedvig.android.feature.chat.di
+
+import com.hedvig.android.feature.chat.ChatEventDataStore
+import com.hedvig.android.feature.chat.ChatEventStore
+import org.koin.dsl.module
+
+val chatModule = module {
+  single<ChatEventStore> { ChatEventDataStore(get()) }
+}
